added modal

+++ user-registration/src/app/registration/registration.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit } from '@angular/core';
-import { User } from '../user';
-import { UserRegistrationService } from '../user-registration.service';
-import { FormBuilder, Validators } from '@angular/forms';
-import { NzMessageService } from 'ng-zorro-antd/message';
+import {Component, OnInit} from '@angular/core';
+import {User} from '../user';
+import {UserRegistrationService} from '../user-registration.service';
+import {FormBuilder, Validators} from '@angular/forms';
+import {NzMessageService} from 'ng-zorro-antd/message';
+import {NzModalService} from 'ng-zorro-antd/modal';
 
 
 @Component({
@@ -14,15 +15,22 @@ export class RegistrationComponent implements OnInit {
 
   users: any;
   email: string;
+  isVisible = false;
 
-  constructor(private service: UserRegistrationService, private formBuilder: FormBuilder, private successMessage: NzMessageService) { }
+  constructor(
+    private service: UserRegistrationService,
+    private formBuilder: FormBuilder,
+    private successMessage: NzMessageService,
+    private modalService: NzModalService
+  ) {
+  }
 
 
   user: User = new User('', '', 0, '');
   message: any;
 
   registrationForm = this.formBuilder.group({
-    name: [this.user.name , Validators.required],
+    name: [this.user.name, Validators.required],
     email: [this.user.email],
     experience: [this.user.experience],
     domain: [this.user.domain]
@@ -41,19 +49,44 @@ export class RegistrationComponent implements OnInit {
       this.successMessage.success('User Successfully Created ', {
         nzDuration: 5000
       });
+      this.ngOnInit();
     });
-    this.ngOnInit();
   }
+
 
   public delete(id: number) {
     if (id != null) {
-      alert('Are u want to sure');
-      const resp = this.service.deleteUser(id);
-      resp.subscribe(data => this.users = data);
+      this.modalService.confirm({
+        nzTitle: 'Are you sure delete this People?',
+        nzOkText: 'Yes',
+        nzOkType: 'danger',
+        nzOnOk: () => {
+          const resp = this.service.deleteUser(id);
+          resp.subscribe(data => this.users = data);
+          this.successMessage.success('People Deleted Successfully', {
+            nzDuration: 3000
+          });
+        },
+        nzOnCancel: () => this.ngOnInit()
+      });
+
     }
   }
 
-  public findByUserEmailid() {
+  showModal(): void {
+    this.isVisible = true;
+  }
+
+  handleOk(): void {
+    console.log('Button ok clicked!');
+    this.isVisible = false;
+  }
+
+  handleCancel(): void {
+    console.log('Button cancel clicked!');
+    this.isVisible = false;
+  }
+  public findByUserEmailId() {
     const resp = this.service.getUsersByEmail(this.email);
     resp.subscribe(data => this.users = data);
   }
